Add seed data and ActiveRecord methods
Added seed data to db/seeds.rb file for all existing models. Also created a new instance method on the User model that takes in a difficulty value as an argument and returns a list of all Tests passed or ever passed at that level. Additionally, added a class method to the Test model that returns an array of names of all Tests sorted in descending order by Category name matching a certain criteria provided as an argument.